Add files via upload
Attempt 2 for LAFD project. I classified incidents into 3 levels of severity based on the incidents' maximum number of dispatch. I hoped to see a difference in the distribution of elapsed_time among incidents of different severity, but there seems to be none. I'm putting the code here just for your reference.